# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Arctis",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development (uvicorn default; see README / docs/Deployment.md)."
    },
    {
      "url": "https://api.example.com",
      "description": "Production \u2014 replace with your public API base URL (docs/Deployment.md)."
    }
  ],
  "paths": {
...
    "/runs/{run_id}": {
      "get": {
        "summary": "Run abrufen",
        "description": "Liefert den gespeicherten Lauf inkl. vollst\u00e4ndigem **execution_summary** (cost, token_usage, steps, evidence, skill_reports). Ghost und Clients holen die Run-ID aus **X-Run-Id** oder **Location** der Customer-Execute-Antwort (POST /customer/workflows/{workflow_id}/execute) und laden Details hier.",
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/customer/workflows/{workflow_id}/execute": {
      "post": {
        "summary": "Execute Workflow Customer Output",
        "operationId": "execute_workflow_customer_output_customer_workflows__workflow_id__execute_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          },
          {
            "$ref": "#/components/parameters/MockHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerExecuteBodySchema"
              },
              "examples": {
                "with_advise_skills": {
                  "summary": "Execute with common advise skills",
                  "description": "Loads reports via GET /runs/{run_id} \u2192 execution_summary.skill_reports.",
                  "value": {
                    "input": {
                      "query": "What is the capital of France?"
                    },
                    "skills": [
                      {
                        "id": "prompt_matrix"
                      },
                      {
                        "id": "routing_explain"
                      },
                      {
                        "id": "cost_token_snapshot"
                      }
                    ]
                  }
                },
                "input_only": {
                  "summary": "Execute without skills",
                  "value": {
                    "input": {
                      "query": "Hello"
                    }
                  }
                },
                "skill_with_params": {
                  "summary": "Skill with params",
                  "value": {
                    "input": {
                      "text": "sample"
                    },
                    "skills": [
                      {
                        "id": "prompt_matrix",
                        "params": {
                          "mode": "advise"
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer Output v1 JSON body. Response headers **X-Run-Id** and **Location** identify the persisted run; clients (e.g. Ghost) load **execution_summary** \u2014 including **skill_reports** \u2014 via **GET /runs/{run_id}** only (body stays minimal; \u00a73.3).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "description": "Customer Output v1: canonical JSON with result and schema_version."
                }
              }
            },
            "headers": {
              "X-Run-Id": {
                "description": "UUID of the persisted Run (same as path param on GET /runs/{run_id}).",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "Location": {
                "description": "Relative URL for GET /runs/{run_id} for this run.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CustomerExecuteBodySchema": {
        "properties": {
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input",
            "description": "Customer input merged with the workflow template before validation and engine run."
          },
          "skills": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CustomerExecuteSkillInvocationSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills",
            "description": "Optional skills to run around the engine execution; results appear under ``execution_summary.skill_reports`` on **GET /runs/{run_id}** (and in evidence)."
          }
        },
        "type": "object",
        "required": [
          "input"
        ],
        "title": "CustomerExecuteBodySchema",
        "description": "JSON body for ``POST /customer/workflows/{workflow_id}/execute``."
      },
      "CustomerExecuteSkillInvocationSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Registered skill id. Unknown ids yield **422** with ``unknown_skill``. Common advise-only ids include ``prompt_matrix``, ``routing_explain``, ``cost_token_snapshot``, ``input_shape``, ``pipeline_config_matrix``, ``evidence_subset``, ``reviewer_explain``."
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params",
            "description": "Optional skill-specific parameters (e.g. ``prompt_matrix`` \u2192 ``{\"mode\": \"advise\"}``)."
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CustomerExecuteSkillInvocationSchema",
        "description": "One entry in the optional ``skills`` array on customer execute."
      },
      "ExecutionSummarySchema": {
        "properties": {
          "cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost",
            "description": "Gesamtkosten in EUR."
          },
          "evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence",
            "description": "E5 Evidence-Envelope (input_evidence, engine_evidence, cost_evidence, \u2026). Bei Customer Execute enth\u00e4lt das Envelope zus\u00e4tzlich **skill_reports** (Spiegel von ``execution_summary.skill_reports``, unver\u00e4ndert). Vollst\u00e4ndiges Bundle \u00fcber GET /runs/{run_id}."
          },
          "token_usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenUsageSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "steps": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Steps",
            "description": "Engine step trace (sofern im Lauf erfasst)."
          },
          "mock": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mock",
            "description": "True wenn der Lauf im E4-Mock-Modus ausgef\u00fchrt wurde."
          },
          "skill_reports": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/SkillReportItemSchema"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Reports",
            "description": "Map skill_id \u2192 Skill-Report (schema_version, payload, provenance). E5: Teil des Evidence-Bundles; identische Daten liegen auch unter ``execution_summary.evidence.skill_reports``. Abruf \u00fcber GET /runs/{run_id}."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ExecutionSummarySchema",
        "description": "Persisted execution metadata: cost, tokens, evidence, steps, mock flag, skill_reports."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunDetailResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input"
          },
          "output": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "pipeline_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Pipeline Version Id"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          },
          "execution_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionSummarySchema"
              },
              {
                "type": "null"
              }
            ],
            "description": "Persistiertes Summary inkl. cost, token_usage, steps, evidence, mock, skill_reports. Nach Customer Execute (POST \u2026/execute) mit Header X-Run-Id / Location hier abrufen."
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "input",
          "pipeline_version_id"
        ],
        "title": "RunDetailResponse",
        "description": "Single run including full execution_summary (cost, tokens, steps, evidence, skill_reports)."
      },
...
      "SkillReportItemSchema": {
        "properties": {
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version",
            "description": "Report format version (e.g. ``1.0``)."
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload",
            "description": "Skill-specific structured output."
          },
          "provenance": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance",
            "description": "Metadata such as ``skill_id``, ``mode: advise``, timestamps."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "SkillReportItemSchema",
        "description": "One skill report as stored under ``execution_summary.skill_reports[<skill_id>]``."
      },
...
      "TokenUsageSchema": {
        "properties": {
          "prompt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "completion": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "type": "object",
        "title": "TokenUsageSchema",
        "description": "Token usage derived from engine cost info (E6)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "parameters": {
      "IdempotencyKeyHeader": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 128
        },
        "description": "Optional. Wiederholbare POST-Requests ohne doppelte Ausf\u00fchrung.\nTenant-scoped. ASCII-only.\n"
      },
      "MockHeader": {
        "name": "X-Arctis-Mock",
        "in": "header",
        "required": false,
        "schema": {
          "type": "boolean"
        },
        "description": "Optional. Erzwingt Mock-Mode (deterministische Engine-Ausf\u00fchrung).\n"
      }
    }
  }
}